requirements.txt
# ...existing dependencies...
django-filter>=23.0
djangorestframework-simplejwt>=5.2.2
django-filter>=23.5